requirement.txt
Django==2.2.5
django-crispy-forms==1.7.2
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
